una idea

+++ libft/strtoint/strtoint.h
@@ -6,7 +6,7 @@
 /*   By: aghergut <aghergut@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 17:35:19 by aghergut          #+#    #+#             */
-/*   Updated: 2025/02/12 18:04:58 by aghergut         ###   ########.fr       */
+/*   Updated: 2025/02/13 16:08:11 by aghergut         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@
 # include <limits.h>
 
 //int to str
-char	*ft_itoa(int n);
+char	*ft_itoa(long n);
 //str to int
-int		ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 
 #endif
